+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version 0.2.0
 *   New parameter types supported: `bool`, `datetime.date`, `datetime.datetime`
 *   `cursor.rowcount` returns number of affected rows for manipulating queries
 *   `Connection` supports `rollback()`
+*   Improved handling of string parameters
 
 
 Version 0.1.0